% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{DimReductionOutput}
\alias{DimReductionOutput}
\title{DimReductionOutput}
\usage{
DimReductionOutput(id)
}
\arguments{
\item{id}{A string representing the namespace ID for the UI component.}
}
\value{
A UI component (tagList) containing the layout for the dimensionality reduction plot and associated input controls.
}
\description{
Creates a UI component for displaying dimensionality reduction plots with options to color by metadata or gene expression.
}
\details{
This function generates a UI layout for displaying a dimensionality reduction plot. It includes:
\itemize{
\item A plot output for the dimensionality reduction plot.
\item A selectize input for choosing the dimensionality reduction method.
\item A selectize input for choosing the coloring method (Nothing, Metadata, or Gene expression).
\item Conditional panels for additional inputs based on the selected coloring method:
\itemize{
\item If "Metadata" is selected, a selectize input for choosing the metadata column.
\item If "Gene expression" is selected, selectize inputs for choosing the gene and the count table.
}
}
}
\examples{
# Example usage in a Shiny app:
# ui <- fluidPage(
#   DimReductionOutput("dim_red_ui")
# )
#
# server <- function(input, output, session) {
#   # Server logic to populate choices and render plot
# }
#
# shinyApp(ui, server)

}
