% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{DimReductionServer}
\alias{DimReductionServer}
\title{DimReductionServer}
\usage{
DimReductionServer(id, db_file = "seurat.duckdb")
}
\arguments{
\item{id}{A string representing the namespace ID for the server module.}

\item{db_file}{A string representing the path to the database file containing the data. Default is "seurat.duckdb".}
}
\value{
A Shiny module server function.
}
\description{
Server logic for handling dimensionality reduction plots in a Shiny app.
}
\details{
This function sets up the server-side logic for a dimensionality reduction plot module in a Shiny app. It performs the following tasks:
\itemize{
\item Reads schema names for count and embedding tables from the database.
\item Updates the choices for the dimensionality reduction method and expression input selectize inputs.
\item Observes changes to the expression input and updates the gene names for coloring by gene expression.
\item Observes changes to the coloring method and updates the metadata names for coloring by metadata.
\item Renders the dimensionality reduction plot based on the selected inputs.
}
}
\examples{
# Example usage in a Shiny app:
# ui <- fluidPage(
#   DimReductionOutput("dim_red_ui")
# )
#
# server <- function(input, output, session) {
#   DimReductionServer("dim_red_ui", db_file = "path/to/database/file")
# }
#
# shinyApp(ui, server)

}
