% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{DotPlot}
\alias{DotPlot}
\title{DotPlot}
\usage{
DotPlot(db_file, gene, split.by = NULL, table = "data")
}
\arguments{
\item{db_file}{A string representing the path to the database file containing gene expression data.}

\item{gene}{A string representing the gene for which the expression data is to be plotted.}

\item{split.by}{An optional string representing the variable to split the y-axis by. Default is NULL.}

\item{table}{A string representing the table name in the database from which to read the gene expression data. Default is "data".}
}
\value{
A ggplot2 object representing the dot plot of gene expression.
}
\description{
Generates a dot plot for gene expression data.
}
\details{
This function reads gene expression data from the specified table in the database, aggregates the data for the dot plot,
and constructs a ggplot object with the specified aesthetics. The dot plot shows the average expression level and the
percentage of cells expressing the gene.
}
