% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{get_default_count_layer}
\alias{get_default_count_layer}
\title{get_default_count_layer}
\usage{
get_default_count_layer(db_file, schema = "layer")
}
\arguments{
\item{db_file}{A string representing the path to the database file.}

\item{schema}{A string representing the schema name to filter tables by. Default is "layer".}
}
\value{
A string representing the name of the first table in the specified schema.
}
\description{
Retrieves the default count layer table name from a specified schema in a database file.
}
\details{
This function retrieves the names of tables within a specified schema from the database file using the \code{get_tables_per_schema_in_db} function. It then filters the tables by the specified schema and returns the name of the first table.
}
\examples{
# Example usage:
# db_file <- system.file("extdata", "pbmc_small.duckdb", package = "lazyseurat")
# get_default_count_layer(db_file)
# get_default_count_layer(db_file, schema = "public")

}
