% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_dim_reduction}
\alias{read_dim_reduction}
\title{read_dim_reduction}
\usage{
read_dim_reduction(db_file, table = "umap")
}
\arguments{
\item{db_file}{A string representing the path to the database file containing dimensionality reduction data.}

\item{table}{A string representing the table name in the database from which to read the dimensionality reduction data. Default is "umap".}
}
\value{
A data frame containing the dimensionality reduction data.
}
\description{
Reads dimensionality reduction data from a specified database file.
}
\details{
This function reads dimensionality reduction data from the specified table in the database file. The data is read using the \code{read_data_with_meta} function.
}
\examples{
# Example usage:
# db_file <- system.file("extdata", "pbmc_small.duckdb", package = "lazyseurat")
# read_dim_reduction(db_file)
# read_dim_reduction(db_file, table = "pca")

}
