% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_gene_expression}
\alias{read_gene_expression}
\title{read_gene_expression}
\usage{
read_gene_expression(db_file, gene = NULL, table = "counts")
}
\arguments{
\item{db_file}{A string representing the path to the database file containing gene expression data.}

\item{gene}{An optional string representing the gene for which the expression data is to be read. Default is NULL.}

\item{table}{A string representing the table name in the database from which to read the gene expression data. Default is "counts".}
}
\value{
A data frame containing the gene expression data.
}
\description{
Reads gene expression data from a specified database file.
}
\details{
This function reads gene expression data from the specified table in the database file. If a gene is specified, only the expression data for that gene is read.
}
\examples{
# Example usage:
# db_file <- system.file("extdata", "pbmc_small.duckdb", package = "lazyseurat")
# read_gene_expression(db_file)
# read_gene_expression(db_file, gene = "MYC")
# read_gene_expression(db_file, gene = "MYC", table = "data")

}
