% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_metadata_names}
\alias{read_metadata_names}
\title{read_metadata_names}
\usage{
read_metadata_names(db_file, max_unique_entries = 10, include_numeric = TRUE)
}
\arguments{
\item{db_file}{A string representing the path to the database file.}

\item{max_unique_entries}{An integer specifying the maximum number of unique entries a column can have to be included. Default is 10.}

\item{include_numeric}{A logical value indicating whether to include numeric columns. Default is TRUE.}
}
\value{
A data frame containing the filtered metadata.
}
\description{
Reads metadata names from a specified database file, filtering columns based on unique entries and optionally excluding numeric columns.
}
\details{
This function reads metadata from the 'metadata' schema in the database file. It filters columns based on the number of unique entries, and optionally excludes numeric columns if \code{include_numeric} is FALSE.
}
\examples{
# Example usage:
# db_file <- system.file("extdata", "pbmc_small.duckdb", package = "lazyseurat")
# read_metadata_names(db_file)
# read_metadata_names(db_file, max_unique_entries = 5)
# read_metadata_names(db_file, include_numeric = FALSE)

}
