% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_schema_names}
\alias{read_schema_names}
\title{read_schema_names}
\usage{
read_schema_names(db_file, schema = "layer")
}
\arguments{
\item{db_file}{A string representing the path to the database file.}

\item{schema}{A string representing the schema name to filter tables by. Default is "layer".}
}
\value{
A character vector containing the names of tables within the specified schema.
}
\description{
Reads the names of tables within a specified schema from a database file.
}
\details{
This function retrieves the names of tables within a specified schema from the database file. It uses the \code{get_tables_per_schema_in_db} function to get all tables and then filters them by the specified schema.
}
\examples{
# Example usage:
# db_file <- system.file("extdata", "pbmc_small.duckdb", package = "lazyseurat")
# read_schema_names(db_file)
# read_schema_names(db_file, schema = "embedding")

}
