% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runLazySeuratShiny}
\alias{runLazySeuratShiny}
\title{runLazySeuratShiny}
\usage{
runLazySeuratShiny(
  db = system.file("extdata", "pbmc_small.duckdb", package = "lazyseurat")
)
}
\arguments{
\item{db}{A string representing the path to the database file containing the Seurat data. Default is "seurat.duckdb".}
}
\value{
A Shiny app object.
}
\description{
Creates a Shiny app template for visualizing Seurat data with dimensionality reduction, violin plots, and dot plots.
}
\details{
This function sets up a Shiny app with the following features:
\itemize{
\item A title panel with the app title.
\item A tabset panel with three tabs:
\itemize{
\item "Dimensionality Reduction" tab containing the DimReductionOutput UI component.
\item "Violin Plot" tab containing the ViolinPlotOutput UI component.
\item "Dot Plot" tab containing the DotPlotOutput UI component.
}
\item Server logic to handle the corresponding server modules for each tab:
\itemize{
\item DimReductionServer for the "Dimensionality Reduction" tab.
\item ViolinPlotServer for the "Violin Plot" tab.
\item DotPlotServer for the "Dot Plot" tab.
}
}
}
\examples{
# Example usage:
# runLazySeuratShiny()

}
