% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ViolinPlotOutput}
\alias{ViolinPlotOutput}
\title{ViolinPlotOutput}
\usage{
ViolinPlotOutput(id)
}
\arguments{
\item{id}{A string representing the namespace ID for the UI component.}
}
\value{
A UI component (tagList) containing the layout for the violin plot and associated input controls.
}
\description{
Creates a UI component for displaying violin plots with options to select count table, gene, and metadata for coloring and splitting.
}
\details{
This function generates a UI layout for displaying a violin plot. It includes:
\itemize{
\item A plot output for the violin plot.
\item A selectize input for choosing the count table.
\item A selectize input for choosing the gene.
\item A selectize input for choosing the metadata column for coloring.
\item A selectize input for choosing the metadata column for separating on the x-axis.
}
}
\examples{
# Example usage in a Shiny app:
# ui <- fluidPage(
#   ViolinPlotOutput("violin_plot_ui")
# )
#
# server <- function(input, output, session) {
#   # Server logic to populate choices and render plot
# }
#
# shinyApp(ui, server)

}
