% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ViolinPlotServer}
\alias{ViolinPlotServer}
\title{ViolinPlotServer}
\usage{
ViolinPlotServer(id, db_file = "seurat.duckdb")
}
\arguments{
\item{id}{A string representing the namespace ID for the server module.}

\item{db_file}{A string representing the path to the database file containing the data. Default is "seurat.duckdb".}
}
\value{
A Shiny module server function.
}
\description{
Server logic for handling violin plots in a Shiny app.
}
\details{
This function sets up the server-side logic for a violin plot module in a Shiny app. It performs the following tasks:
\itemize{
\item Reads schema names for count tables from the database.
\item Updates the choices for the count table and gene selectize inputs.
\item Reads metadata names and updates the choices for coloring and splitting the violin plot.
\item Renders the violin plot based on the selected inputs.
}
}
\examples{
# Example usage in a Shiny app:
# ui <- fluidPage(
#   ViolinPlotOutput("violin_plot_ui")
# )
#
# server <- function(input, output, session) {
#   ViolinPlotServer("violin_plot_ui", db_file = "path/to/database/file")
# }
#
# shinyApp(ui, server)

}
