% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{VlnPlot}
\alias{VlnPlot}
\title{VlnPlot}
\usage{
VlnPlot(db_file, gene, x = NULL, split.by = NULL, table = "counts")
}
\arguments{
\item{db_file}{A string representing the path to the database file containing gene expression data.}

\item{gene}{A string representing the gene for which the expression data is to be plotted.}

\item{x}{(Optional) A string representing the variable to be used for the x-axis. Default is NULL.}

\item{split.by}{(Optional) A string representing the variable to be used for splitting the fill color. Default is NULL.}

\item{table}{A string representing the table to use for gene expression data. Default is "counts".}
}
\value{
A ggplot2 object representing the violin plot of gene expression.
}
\description{
Generates a violin plot for gene expression data.
}
\examples{
# Example usage:
# db_file <- "path/to/database/file"
# gene <- "GeneName"
# VlnPlot(db_file, gene)
# VlnPlot(db_file, gene, x = "CellType")
# VlnPlot(db_file, gene, x = "CellType", split.by = "Condition")

}
